Refactoring home screen, starting a new page and fixing the popup menus
